the id should be a long, not an int

+++ web_module/branches/studio_toad_merge_2008_04_17/src/main/com/topcoder/web/common/dao/PathDAO.java
@@ -10,5 +10,5 @@ import com.topcoder.web.common.model.Path;
  * @version $Id$
  *          Create Date: Apr 21, 2008
  */
-public interface PathDAO extends GenericDAO<Path, Integer> {
+public interface PathDAO extends GenericDAO<Path, Long> {
 }
